+++ jpa/src/main/java/vezh_bank/config/JpaConfig.java
@@ -1,26 +1,44 @@
 package vezh_bank.config;
 
-import org.springframework.context.annotation.*;
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Scope;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
-import vezh_bank.persistence.DataBaseService;
-import vezh_bank.persistence.dao.*;
-import vezh_bank.persistence.dao.impl.*;
 
 import javax.persistence.EntityManagerFactory;
+import javax.sql.DataSource;
 
 @Configuration
 @EnableTransactionManagement
-@ComponentScan(basePackageClasses = DataBaseService.class)
+//@ComponentScan(basePackageClasses = DataBaseService.class)
 public class JpaConfig {
 
     @Bean
+    @Primary
+    public DataSource dataSource() {
+        return DataSourceBuilder.create()
+                .driverClassName("com.mysql.cj.jdbc.Driver")
+                .url("jdbc:mysql://localhost:3306/vezh_bank?" +
+                        "useUnicode=true&" +
+                        "useJDBCCompliantTimezoneShift=true&" +
+                        "useLegacyDatetimeCode=false&" +
+                        "serverTimezone=UTC")
+                .username("root")
+                .password("root")
+                .build();
+    }
+
+    @Bean
     @Scope(value = "singleton")
-    public EntityManagerFactory emf() {
+    public EntityManagerFactory emf(DataSource dataSource) {
         LocalContainerEntityManagerFactoryBean factory = new LocalContainerEntityManagerFactoryBean();
+        factory.setDataSource(dataSource);
         factory.setPersistenceUnitName("vezh_bank");
         factory.setJpaVendorAdapter(new HibernateJpaVendorAdapter());
         factory.afterPropertiesSet();
@@ -30,47 +48,5 @@ public class JpaConfig {
     @Bean
     public PlatformTransactionManager txMgr(EntityManagerFactory emf) {
         return new JpaTransactionManager(emf);
-    }
-
-    @Bean
-    @Lazy
-    public CardDao cardDao() {
-        return new CardDaoImpl(){};
-    }
-
-    @Bean
-    @Lazy
-    public CurrencyDao currencyDao() {
-        return new CurrencyDaoImpl(){};
-    }
-
-    @Bean
-    @Lazy
-    public EventDao eventDao() {
-        return new EventDaoImpl(){};
-    }
-
-    @Bean
-    @Lazy
-    public RoleDao roleDao() {
-        return new RoleDaoImpl(){};
-    }
-
-    @Bean
-    @Lazy
-    public TransactionDao transactionDao() {
-        return new TransactionDaoImpl(){};
-    }
-
-    @Bean
-    @Lazy
-    public UserDao userDao() {
-        return new UserDaoImpl(){};
-    }
-
-    @Bean
-    @Lazy
-    public UserRequestDao userRequestDao() {
-        return new UserRequestDaoImpl(){};
     }
 }
