+++ jpa/src/main/java/vezh_bank/persistence/entity/UserRequest.java
@@ -15,7 +15,7 @@ public class UserRequest {
 
     @ManyToOne
     @JoinTable(name = "USERS")
-    @Column(name = "USER_ID")
+//    @Column(name = "USER_ID")
     private User user;
 
     @Column(name = "CREATION_DATE")
