+++ jpa/src/main/java/vezh_bank/persistence/entity/UserRole.java
@@ -3,7 +3,6 @@ package vezh_bank.persistence.entity;
 import vezh_bank.enums.Role;
 
 import javax.persistence.*;
-import java.util.List;
 
 @Entity
 @Table(name = "USER_ROLES")
@@ -16,8 +15,8 @@ public class UserRole {
     @Column(name = "ROLE_NAME")
     private String name;
 
-    @OneToMany(mappedBy = "role")
-    private List<User> users;
+//    @OneToMany(mappedBy = "role")
+//    private List<User> users;
 
     public UserRole() {
     }
