+++ pom.xml
@@ -34,6 +34,7 @@
             <artifactId>log4j</artifactId>
             <version>1.2.17</version>
         </dependency>
+
     </dependencies>
 
     <build>
