+++ vezh_bank/src/main/java/core/config/VezhBankConfiguration.java
@@ -3,16 +3,20 @@ package core.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.web.servlet.ViewResolver;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.view.InternalResourceViewResolver;
 import vezh_bank.config.JpaConfig;
+import vezh_bank.persistence.DataBaseService;
+import vezh_bank.persistence.dao.*;
+import vezh_bank.persistence.dao.impl.*;
 
 @Configuration
 @EnableWebMvc
-@ComponentScan(basePackageClasses = JpaConfig.class)
+@ComponentScan(basePackageClasses = {JpaConfig.class, DataBaseService.class})
 public class VezhBankConfiguration implements WebMvcConfigurer {
 
     @Override
@@ -25,6 +29,48 @@ public class VezhBankConfiguration implements WebMvcConfigurer {
         InternalResourceViewResolver internalResourceViewResolver = new InternalResourceViewResolver("/WEB-INF/views/", ".jsp");
 //        internalResourceViewResolver.setRedirectHttp10Compatible(false);
         return internalResourceViewResolver;
+    }
+
+    @Bean
+    @Lazy
+    public CardDao cardDao() {
+        return new CardDaoImpl(){};
+    }
+
+    @Bean
+    @Lazy
+    public CurrencyDao currencyDao() {
+        return new CurrencyDaoImpl(){};
+    }
+
+    @Bean
+    @Lazy
+    public EventDao eventDao() {
+        return new EventDaoImpl(){};
+    }
+
+    @Bean
+    @Lazy
+    public RoleDao roleDao() {
+        return new RoleDaoImpl(){};
+    }
+
+    @Bean
+    @Lazy
+    public TransactionDao transactionDao() {
+        return new TransactionDaoImpl(){};
+    }
+
+    @Bean
+    @Lazy
+    public UserDao userDao() {
+        return new UserDaoImpl(){};
+    }
+
+    @Bean
+    @Lazy
+    public UserRequestDao userRequestDao() {
+        return new UserRequestDaoImpl(){};
     }
 
 }
